+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Version changes are pinned to SDK releases.
 
 ## [Unreleased]
 
+## [1.23.3]
+
+- New function closeAllPositions(). ([#362](https://github.com/zetamarkets/sdk/pull/362))
+- New option to specify retry amount for txs. Use CrossClient.setTxRetryAmount(). ([#362](https://github.com/zetamarkets/sdk/pull/362))
+
 ## [1.23.2]
 
 - Real-time maker rebate. ([#359](https://github.com/zetamarkets/sdk/pull/359))